split `session`
