+++ plugins/platform-core/src/session.ts
@@ -21,19 +21,9 @@ import { generateId } from './objectid'
 
 import core, { Instance, Type, Session, Values, Adapter, Cursor, CoreDomain } from '.'
 
+import { Prototypes, createPrototypes } from './instance'
+
 export function createSession (platform: Platform, modelDb: MemDb, coreProtocol: CoreProtocol, broadcastXact: (info: CommitInfo, originator?: Session) => void, closeSession: (session: Session) => void): Session {
-
-  type Konstructor<T extends Obj> = new (obj: Layout<Obj>) => Instance<T>
-
-  enum Stereotype {
-    EMB,
-    DOC
-  }
-
-  // M O D E L
-
-  const konstructors = new Map<Ref<Class<Obj>>, Konstructor<Obj>>()
-  const prototypes = new Map<Ref<Class<Obj>>, Object>()
 
   const created = new Map<Ref<Doc>, Instance<Doc>>()
   const updated = new Map<Ref<Doc>, Instance<Doc>>()
@@ -43,135 +33,15 @@ export function createSession (platform: Platform, modelDb: MemDb, coreProtocol:
     getSession: (): Session => session
   }
 
-  async function getPrototype<T extends Obj> (_class: Ref<Class<T>>, stereotype: Stereotype): Promise<Object> {
-    const prototype = prototypes.get(_class)
-    if (prototype) { return prototype }
-
-    const clazz = modelDb.get(_class) as Layout<Class<Doc>>
-    const parent = clazz._extends ? await getPrototype(clazz._extends as string as Ref<Class<Obj>>, stereotype) : CoreRoot
-    const proto = Object.create(parent)
-    prototypes.set(_class, proto)
-
-    if (clazz._native) {
-      const native = await platform.getResource(clazz._native as unknown as Resource<Object>) // TODO: must `resolve`! we need to have getPrototype async for this.
-      if (!native) { throw new Error(`something went wrong, can't load '${clazz._native}' resource`) }
-      const descriptors = Object.getOwnPropertyDescriptors(native)
-      Object.defineProperties(proto, descriptors)
-      // prototypes.set(_class, proto)
-      return proto
-    }
-
-    const attributes = clazz._attributes
-    for (const key in attributes) {
-      if (key === '_default') { continue } // we do not define `_default`'s type, it's infinitevely recursive :)
-      const attr = attributes[key]
-      // console.log(attr)
-      const attrInstance = await instantiateEmb(attr) as Instance<Type<any>>
-      // console.log(attrInstance)
-
-      // if (typeof attrInstance.exert !== 'function') {
-      //   throw new Error('exert is not a function')
-      // }
-
-      const exertFactory = attrInstance.exert
-
-      if (typeof exertFactory !== 'function') {
-        throw new Error('exertFactory is not a function ' + attrInstance._class)
-      }
-
-      const exert = await exertFactory.call(attrInstance)
-      if (typeof exert !== 'function') {
-        throw new Error('exert is not a function')
-      }
-
-      const hibernate = attrInstance.hibernate
-      if (typeof hibernate !== 'function') {
-        throw new Error('hibernate is not a function')
-      }
-      const hibernateBound = hibernate.bind(attrInstance)
-
-      const fullKey = stereotype === Stereotype.DOC ?
-        key.startsWith('_') ? key : attributeKey(_class, key) :
-        key
-
-      Object.defineProperty(proto, key, {
-        get (this: Instance<Obj>) {
-          let value = (this.__update as any)[fullKey]
-          if (!value) {
-            value = (this.__layout as any)[fullKey]
-          }
-          return exert(value, this.__layout, key)
-        },
-        set (this: Instance<Obj>, value: any) {
-          (this.__update as any)[fullKey] = hibernateBound(value)
-          const id = (this.__layout as Layout<Doc>)._id
-          if (id) { updated.set(id, this as Instance<Doc>) }
-        },
-        enumerable: true
-      })
-    }
-    return proto
-  }
-
-  async function getKonstructor<T extends Obj> (_class: Ref<Class<T>>, stereotype: Stereotype): Promise<Konstructor<T>> {
-    const konstructor = konstructors.get(_class)
-    if (konstructor) { return konstructor as unknown as Konstructor<T> }
-    else {
-      const proto = await getPrototype(_class, stereotype)
-      const ctor = {
-        [_class]: function (this: Instance<Obj>, obj: Layout<Doc>) {
-          this.__layout = obj
-          this.__update = {} as Layout<Doc>
-        }
-      }[_class] // A trick to `name` function as `_class` value
-      proto.constructor = ctor
-      ctor.prototype = proto
-      konstructors.set(_class, ctor as unknown as Konstructor<Obj>)
-      return ctor as unknown as Konstructor<T>
-    }
-  }
-
-  async function instantiateEmb<T extends Emb> (obj: Layout<Obj>): Promise<Instance<T>> {
-    const ctor = await getKonstructor(obj._class, Stereotype.EMB)
-    return new ctor(obj) as Instance<T>
-  }
-
-  async function instantiateDoc<T extends Doc> (obj: Layout<Doc>): Promise<Instance<T>> {
-    const ctor = await getKonstructor(obj._class, Stereotype.DOC)
-    return new ctor(obj) as unknown as Instance<T>
-  }
-
-  // TODO: employ prototypes to get rid of this
-  // async function getAttribute (_class: Ref<Class<Obj>>, key: string) {
-  //   let cls = _class as Ref<Classifier<Obj>> | undefined
-  //   console.log('get attribute: ' + key + ' ' + cls)
-  //   while (cls) {
-  //     const clazz = await getInstance(core.class.Class, cls)
-  //     console.log('instance: ' + clazz)
-  //     const attributes = clazz._attributes as { [key: string]: Instance<Type<any>> | undefined }
-  //     const attr = attributes[key]
-  //     if (attr) { return attr }
-  //     cls = clazz._extends
-  //     console.log(clazz._extends)
-  //   }
-  //   return undefined
-  // }
+  const prototypes = createPrototypes(platform, modelDb, CoreRoot, (doc: Instance<Doc>) => {
+    updated.set(doc._id, doc)
+  })
 
   async function newInstance<M extends Doc> (_class: Ref<Class<M>>, values: Values<Omit<M, keyof Doc>>, id?: Ref<M>): Promise<Instance<M>> {
     const _id = id ?? generateId() as Ref<Doc>
     const layout = { _class, _id } as Layout<Doc>
 
-    // for (const key in values) {
-    //   const attr = await getAttribute(_class, key)
-    //   if (!attr) { throw new Error('attribute not found') }
-    //   const value = (values as any)[key]
-    //   const h = attr.hibernate
-    //   if (!h) { throw new Error('no hibernate function') }
-    //   const lv = h.bind(attr)(value)
-    //     ; (layout as any)[key] = lv
-    // }
-
-    const instance = await instantiateDoc(layout)
+    const instance = await prototypes.instantiateDoc(layout)
     for (const key in values) {
       (instance as any)[key] = (values as any)[key]
     }
@@ -206,20 +76,7 @@ export function createSession (platform: Platform, modelDb: MemDb, coreProtocol:
     }
 
     if (!doc) { throw new Error('no document ,id: ' + _id) }
-    return instantiateDoc(doc as T)
-  }
-
-  async function as<T extends Doc, A extends Doc> (doc: Instance<T>, _class: Ref<Class<A>>): Promise<Instance<A>> {
-    if (!is(doc, _class)) {
-      console.log('Warning:' + _class + ' instance does not mixed into `' + doc._class + '`')
-    }
-    const ctor = await getKonstructor(_class, Stereotype.DOC)
-    return new ctor(doc.__layout as unknown as A)
-  }
-
-  function is<T extends Doc, M extends Doc> (doc: Instance<T>, _class: Ref<Class<M>>): boolean {
-    const mixins = doc._mixins as Ref<Class<Doc>>[]
-    return mixins && mixins.includes(_class as Ref<Class<Doc>>)
+    return prototypes.instantiateDoc(doc as T)
   }
 
   // C O M M I T S
@@ -262,7 +119,7 @@ export function createSession (platform: Platform, modelDb: MemDb, coreProtocol:
       if (find.length > 0) {
         console.log('match found: ' + find.length)
         // q.result.push(...find)
-        const foundInstances = await Promise.all(find.map(doc => instantiateDoc(doc)))
+        const foundInstances = await Promise.all(find.map(doc => prototypes.instantiateDoc(doc)))
         q.instances.push(...foundInstances)
         q.listener(q.instances)
       }
@@ -301,7 +158,7 @@ export function createSession (platform: Platform, modelDb: MemDb, coreProtocol:
         await Promise.all(promises)
         console.log('LAYOUT: ', layout)
         const layouts = coreProtocol.find(_class as Ref<Class<Doc>>, layout)
-        const result = await layouts.then(resolved => (resolved as T[]).map(doc => instantiateDoc(doc)))
+        const result = await layouts.then(resolved => (resolved as T[]).map(doc => prototypes.instantiateDoc(doc)))
         return Promise.all(result as unknown as Promise<Instance<T>>[])
       }
     }
@@ -329,7 +186,7 @@ export function createSession (platform: Platform, modelDb: MemDb, coreProtocol:
 
     async function getFromCache (): Promise<void> {
       return (await cache).find(_class as Ref<Class<Doc>>, query as unknown as AnyLayout) // !!!!! WRONG, need hibernate
-        .then(result => Promise.all(result.map(doc => instantiateDoc(doc))))
+        .then(result => Promise.all(result.map(doc => prototypes.instantiateDoc(doc))))
         .then(result => {
           if (!done) {
             q.instances = result
@@ -344,7 +201,7 @@ export function createSession (platform: Platform, modelDb: MemDb, coreProtocol:
       .then(async result => {
         done = true
           ; ((await cache).cache(result)).then(() => console.log('RESULTS CACHED'))
-        return Promise.all(result.map(doc => instantiateDoc(doc)))
+        return Promise.all(result.map(doc => prototypes.instantiateDoc(doc)))
       })
       .then(async result => {
         q.instances = result
@@ -396,15 +253,12 @@ export function createSession (platform: Platform, modelDb: MemDb, coreProtocol:
     getClassHierarchy,
     newInstance,
     getInstance,
-    as,
-    is,
     find,
     query,
     delete: del,
     adapt,
-    instantiateEmb,
-    getPrototype,
-    getModel () { return modelDb }
+    getModel () { return modelDb },
+    ...prototypes
   }
 
   return session
